git-p4: Handle release branch branching strategy in perforce
In perforce, the mapping of a branch can be updated after that branch is
committed. This means that when we initially try to import a branch, its
parent branch may not yet exist in the git repo. We can workaround this
by recursing through its parents parents until we find one that already
exists

Signed-off-by: Paul Becotte <pjbecotte@hey.com>